+++ frontend/app/src/components/history/events/HistoryEventsAction.vue
@@ -16,6 +16,7 @@ const emit = defineEmits<{
   (e: 'toggle-ignore', event: HistoryEventEntry): void;
   (e: 'redecode', data: EvmChainAndTxHash): void;
   (e: 'reset', event: EvmHistoryEvent): void;
+  (e: 'delete-tx', data: EvmChainAndTxHash): void;
 }>();
 
 const { event } = toRefs(props);
@@ -28,6 +29,7 @@ const addEvent = (event: HistoryEventEntry) => emit('add-event', event);
 const toggleIgnore = (event: HistoryEventEntry) => emit('toggle-ignore', event);
 const redecode = (data: EvmChainAndTxHash) => emit('redecode', data);
 const resetEvent = (event: EvmHistoryEvent) => emit('reset', event);
+const deleteTxAndEvents = (data: EvmChainAndTxHash) => emit('delete-tx', data);
 </script>
 
 <template>
@@ -57,10 +59,7 @@ const resetEvent = (event: EvmHistoryEvent) => emit('reset', event);
           @click="addEvent(event)"
         >
           <template #prepend>
-            <RuiIcon
-              class="text-rui-text-secondary"
-              name="add-line"
-            />
+            <RuiIcon name="add-line" />
           </template>
           {{ t('transactions.actions.add_event_here') }}
         </RuiButton>
@@ -69,10 +68,7 @@ const resetEvent = (event: EvmHistoryEvent) => emit('reset', event);
           @click="toggleIgnore(event)"
         >
           <template #prepend>
-            <RuiIcon
-              class="text-rui-text-secondary"
-              :name="event.ignoredInAccounting ? 'eye-line' : 'eye-off-line'"
-            />
+            <RuiIcon :name="event.ignoredInAccounting ? 'eye-line' : 'eye-off-line'" />
           </template>
           {{
             event.ignoredInAccounting
@@ -87,10 +83,7 @@ const resetEvent = (event: EvmHistoryEvent) => emit('reset', event);
             @click="redecode(toEvmChainAndTxHash(evmEvent))"
           >
             <template #prepend>
-              <RuiIcon
-                class="text-rui-text-secondary"
-                name="restart-line"
-              />
+              <RuiIcon name="restart-line" />
             </template>
             {{ t('transactions.actions.redecode_events') }}
           </RuiButton>
@@ -100,12 +93,20 @@ const resetEvent = (event: EvmHistoryEvent) => emit('reset', event);
             @click="resetEvent(evmEvent)"
           >
             <template #prepend>
-              <RuiIcon
-                class="text-rui-text-secondary"
-                name="file-edit-line"
-              />
+              <RuiIcon name="file-edit-line" />
             </template>
             {{ t('transactions.actions.reset_customized_events') }}
+          </RuiButton>
+          <RuiButton
+            variant="list"
+            color="error"
+            :disabled="loading"
+            @click="deleteTxAndEvents(toEvmChainAndTxHash(evmEvent))"
+          >
+            <template #prepend>
+              <RuiIcon name="delete-bin-line" />
+            </template>
+            {{ t('transactions.actions.delete_transaction') }}
           </RuiButton>
         </template>
       </div>